Add files via upload
This is the version 3 upload of the user profile page.

Here are additions since the previous version:
- Expand upon userProfile JavaScript so that it incorporates applies to all site questions and updates web page with user inputs.
- Add progress bar to user profile.
- Slightly tweaked JavaScript.

It turns out that I have misunderstood the requirements for this user profile. I have bascially had 10 questions total on screen, but seperated by 3 sections all on screen at once. The way i was supposed to do this user profile was to show the user each section individually with at least 3 questions per section. This is something I aim to rectify in future iterations in a REWORK.

The good side, is that i have a better idea of how to tackle the JavaScript from the progress of these last 3 uploads, so should hopefully speed up the process of reworking the userProfile webpage.

To do:
- REWORK  userProfile webpage, so it matches requirements.
- Add the ability to edit answers to questions.
- Use arrays to store profile questions and user answers.
- Incorporate navBar and site logo.
- Create and link a page editor webpage.
- Add a link to the websites home page.
- Add a background image.
- Tweak userProfile CSS file as required to give the website a cohesive look and feel.
- Adjust links to webpages so they link to ones uploaded on server.