Add files via upload
This is the version 5 upload of the userProfile.

Here are the additions since the previous version:
- Implemented use of section 2 and section 3 in the 'answerQuestions' function.
- Implemented input validation for all sections, for inputs that require them. (partial)
- Implemented input sanitisation. (partial)
- Created function to append valid inputs to array and to update HTML page/file.
- Made sure all sections are presented together at the end.
- Added a button to Edit Answers.

______________________________________________________________________
In this version, I have made a lot of progress since the last upload, with the biggest leaps being the implementation of some input validation, implementing sections 2 and 3, as well as displaying all sections together. However, I have added (partial) to the end of the sections on input validation and input sanitisation.

This is because the input validation is not very comprehensive as I would like. For example, the answer to the question 'phone number' a user input would be accepted if it contains a number, even if there are letters in the input, which phone numbers do not have. This is because I was struggling to implement scanning inputs to only be numbers for example, even though it seems like an easy task.

I think because of the large advancement in JavaScript in particular, alongside my code being by no means the most efficient or user friendly, caused me to struggle to undestand my own code. This is where I need to implement effective code comments.
______________________________________________________________________

To do:
- Add a the ability to answer unanswered questions. - (currently Work In Progress in this version, reliable functionality not achieved).
- Add a function to edit answers if required. - (currently Work In Progress in this version, reliable functionality not achieved).
- Improve upon input validation? - (if possible)
- Improve upon input sanitisation? - (if possible)
- Add meaningful code comments.
- Add a link to a 'page editor' webpage.
- Add a link to the websites 'home page'.
- Add navBar to webpage.
- Add website logo to webpage.
- Update CSS to make website seem cohesive; e.g. adding background image etc.