Add files via upload
This is the version 7 upload of the userProfile.

Here are the additions I have made since the previous version:
- Added a link to a 'page editor' webpage.
- Added a link to the websites 'home page'.
- Added the navBar to webpage.
- Added the website logo to webpage.
- Updated CSS to make website seem cohesive; e.g. adding background image etc.

This verison is the one that I would consider final. It has completed all the milestones I set for it during its development. 

If there was something I could improve in future versions, it would be the input validation and sanitisation, so it works more thoroughly with inputs. but this is not possible at the moment due to project time constraints.