Add files via upload
This is the version 2 upload of the teamPage.

Since the previous version, I have:
- Incorporated navbar.
- Incorporated logo.
- Incorporated page editor link.
- Added backgroud image.
- Improved CSS in an attempt to make the webpage feel cohesive to the website.

This webpage is in my opinion considered complete / final.

If there was anything I would change if I re-did this section, it would be to put the title 'Team page' in within and inset border neatly, so it mirrors the design of my other webpage 'userProfile'. However, due to project time constrains, this is not possible.