Add files via upload
This is the version 8 upload of the userProfile.

Uploaded this new version, which passes validation.

This verison is the one that I would consider final. It has completed all the milestones I set for it during its development. 

If there was something I could improve in future versions, it would be the input validation and sanitisation, so it works more thoroughly with inputs. but this is not possible at the moment due to project time constraints.